Simplify
Add the ability to do ?amt=xxx.

After trying this with large amounts of data, it turns out
that serde_json is faster than the hand crafted json from previous
commits.  While hand-crafted seems quicker for 1000 users, it's not
worth the extra maintenance especially if larger values are slower
with hand crafted json.

If we just want to win benchmarking at 1000, use the previous commit.
But this is better code.